Fix corner-case event count doubling
When performing a DroneBL query, and the comment resembles a comment we
would have added (it contains an event count), and we have not yet updated
our local event count with this figure (we're querying for an address that
was added before we were (re)started), we add the existing count to our
current count to avoid losing track of the number of events.

Unfortunately we also do this when performing a first-time update for a
listing that wasn't originally submitted by us. This results in the
following unfortunate sequence of events:

  - An IP address has been previously listed outside of this script and
    thus does not have an event count

  - We observe an abuse event from that IP address (even once)

  - That information is recorded in the list of observed IP addresses

  - On the next DroneBL query, it is returned in the results, but does
    not have an event count, so we do nothing to the local count

  - We submit an update request for the listing, with our local count

  - On the next DroneBL query, it is returned in the results, has an
    event count, and we erroneously add it to our local count

In effect, this results in a doubling (or more) of the event count.

Fix it.

Also, correct an erroneous comment below that block.