Support client reconnections to the IRC server
If we are disconnected from our IRC server, we were relying on the
while loop at the beginning of our Tasks to eventually bail out,
ending all running Tasks and eventually the program itself.

This presents a data loss scenario, in which events observed after
the last submission interval but before the disconnection are lost.

Now we create Tasks explicitly (instead of using asyncio to gather
them), and the loops always run. When the client connects, the
Tasks are scheduled, and when it disconnects, they are cancelled.

Another benefit to this approach is that the SIGTERM handler can
also immediately cancel the Tasks, therefore resulting in the
program itself exiting in short order (rather than waiting for the
last Task's while loop to notice we're not connected anymore).

This also means we no longer need to be run as an asyncio program
(i.e. main() doesn't have to be async), because main() no longer
indirectly creates Tasks. We also don't have to grab our own event
loop anymore, relying on Pydle to do that for us.