Forget about IP addresses after successfully adding or updating them
Remebering them for the lifetime of the process prevents incidents
from being delisted; the bot would just re-submit them on the next
interval. This means someone would have to restart the bot before
removing an incident, which is less than optimal.

We can rely on the existing backdating code to properly restore the
first-seen and event count data for each IP address when we perform
the status query before each new submission.

This also means that the IP address stats reporting function is now
not useful, as it will only show events since the last successful
submission, so remove that too.